disk: enable controller on Windows
