+++ pkg/disk/controllerserver.go
@@ -1,5 +1,3 @@
-//go:build !windows
-
 /*
 Copyright 2019 The Kubernetes Authors.
 
@@ -361,7 +359,7 @@ func (cs *controllerServer) ControllerPublishVolume(ctx context.Context, req *cs
 
 	klog.Infof("ControllerPublishVolume: start attach disk: %s to node: %s", req.VolumeId, req.NodeId)
 
-	r, err := cs.ad.attachDisk(ctx, req.VolumeId, req.NodeId, false)
+	r, err := cs.ad.attachDisk(ctx, req.VolumeId, req.NodeId)
 	if err != nil {
 		klog.Errorf("ControllerPublishVolume: attach disk: %s to node: %s with error: %s", req.VolumeId, req.NodeId, err.Error())
 		return nil, err
