+++ pkg/disk/datacache/data_cache.go
@@ -14,58 +14,9 @@ import (
 
 	utilsio "github.com/kubernetes-sigs/alibaba-cloud-csi-driver/pkg/utils/io"
 	"golang.org/x/sys/unix"
-	"k8s.io/apimachinery/pkg/api/resource"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/klog/v2"
 )
-
-type Mode string
-
-const (
-	Writethrough Mode = "writethrough"
-	Writeback    Mode = "writeback"
-)
-
-type Opts struct {
-	Size resource.Quantity
-	Mode Mode
-}
-
-func (o *Opts) Enabled() bool {
-	return !o.Size.IsZero()
-}
-
-const (
-	modeKey = "dataCacheMode"
-	sizeKey = "dataCacheSize"
-)
-
-func GetOpts(opts map[string]string, d *Opts) error {
-	if s := opts[sizeKey]; s != "" {
-		size, err := resource.ParseQuantity(s)
-		if err != nil {
-			return fmt.Errorf("invalid %s: %w", sizeKey, err)
-		}
-		d.Size = size
-	}
-
-	switch m := Mode(opts[modeKey]); m {
-	case "", Writeback, Writethrough:
-		d.Mode = m
-	default:
-		return fmt.Errorf("unrecognized %s: %s", modeKey, m)
-	}
-
-	if d.Mode != "" || !d.Size.IsZero() {
-		if d.Size.IsZero() {
-			return fmt.Errorf("must specify non-zero %s for dataCache", sizeKey)
-		}
-		if d.Mode == "" {
-			d.Mode = Writethrough
-		}
-	}
-	return nil
-}
 
 func loggedClose(logger klog.Logger, fd int) {
 	if err := unix.Close(fd); err != nil {
