+++ pkg/disk/datacache/data_cache_unsupported.go
@@ -1,4 +1,4 @@
-//go:build !linux
+//go:build !linux && !windows
 
 package datacache
