+++ pkg/disk/disk.go
@@ -1,5 +1,3 @@
-//go:build !windows
-
 /*
 Copyright 2019 The Kubernetes Authors.
 
@@ -101,7 +99,6 @@ func NewServers(m metadata.MetadataProvider, ecsV2 cloud.ECSv2Interface, endpoin
 
 	if serviceType&utils.Node != 0 {
 		GlobalConfigVar.NodeID = metadata.MustGet(m, metadata.InstanceID)
-		DefaultDeviceManager.DisableSerial = IsVFNode()
 	} else {
 		GlobalConfigVar.NodeID = "not-retrieved" // make csi-common happy
 	}
@@ -121,7 +118,7 @@ func NewServers(m metadata.MetadataProvider, ecsV2 cloud.ECSv2Interface, endpoin
 		servers.ControllerServer = NewControllerServer(csiCfg, client, ecsV2, m)
 	}
 	if serviceType&utils.Node != 0 {
-		servers.NodeServer = NewNodeServer(client, ecsV2, m, useLabeler)
+		servers.NodeServer = NewNodeServer(csiCfg, client, ecsV2, m, useLabeler)
 	}
 	if features.FunctionalMutableFeatureGate.Enabled(features.EnableVolumeGroupSnapshots) {
 		servers.GroupControllerServer = NewGroupControllerServer()
@@ -209,7 +206,6 @@ func GlobalConfigSet(m metadata.MetadataProvider, csiCfg utils.Config) {
 	} else {
 		RegionalDiskTopologyKey = v1.LabelTopologyRegion
 	}
-	DefaultDeviceManager.EnableDiskPartition = csiCfg.GetBool("disk-partition-enable", "DISK_PARTITION_ENABLE", true)
 	klog.Infof("Starting with GlobalConfigVar: ADControllerEnable(%t), DiskTagEnable(%t), DiskBdfEnable(%t), MetricEnable(%t), DetachDisabled(%t), DetachBeforeDelete(%t), ClusterID(%s)",
 		GlobalConfigVar.ADControllerEnable,
 		GlobalConfigVar.DiskTagEnable,
