+++ pkg/disk/group_volume_snapshot_utils_test.go
@@ -1,5 +1,3 @@
-//go:build !windows
-
 package disk
 
 import (
