+++ pkg/disk/groupcontrollerserver.go
@@ -1,5 +1,3 @@
-//go:build !windows
-
 package disk
 
 import (
