+++ pkg/disk/utils.go
@@ -1,5 +1,3 @@
-//go:build !windows
-
 /*
 Copyright 2019 The Kubernetes Authors.
 
@@ -42,25 +40,20 @@ import (
 	aliyunep "github.com/aliyun/alibaba-cloud-sdk-go/sdk/endpoints"
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/ecs"
 	"github.com/container-storage-interface/spec/lib/go/csi"
-	"github.com/containerd/ttrpc"
 	volumeSnapshotV1 "github.com/kubernetes-csi/external-snapshotter/client/v8/apis/volumesnapshot/v1"
 	snapClientset "github.com/kubernetes-csi/external-snapshotter/client/v8/clientset/versioned"
 	"github.com/kubernetes-sigs/alibaba-cloud-csi-driver/pkg/cloud"
 	"github.com/kubernetes-sigs/alibaba-cloud-csi-driver/pkg/cloud/metadata"
 	"github.com/kubernetes-sigs/alibaba-cloud-csi-driver/pkg/common"
 	"github.com/kubernetes-sigs/alibaba-cloud-csi-driver/pkg/disk/datacache"
-	proto "github.com/kubernetes-sigs/alibaba-cloud-csi-driver/pkg/disk/proto"
 	"github.com/kubernetes-sigs/alibaba-cloud-csi-driver/pkg/utils"
 	utilshttp "github.com/kubernetes-sigs/alibaba-cloud-csi-driver/pkg/utils/http"
-	utilsio "github.com/kubernetes-sigs/alibaba-cloud-csi-driver/pkg/utils/io"
 	"github.com/kubernetes-sigs/alibaba-cloud-csi-driver/pkg/version"
-	perrors "github.com/pkg/errors"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/client-go/tools/record"
 	"k8s.io/klog/v2"
-	k8smount "k8s.io/mount-utils"
 	"k8s.io/utils/ptr"
 )
 
@@ -72,6 +65,10 @@ var (
 
 const (
 	instanceTypeInfoAnnotation = "alibabacloud.com/instance-type-info"
+	// PVC annotation key of KMS key ID, override the storage class parameter kmsKeyId
+	KMSKeyID = "alibabacloud.com/kms-key-id"
+	// CreateDiskARN ARN parameter of the CreateDisk interface
+	CreateDiskARN = "alibabacloud.com/createdisk-arn"
 )
 
 // RoleAuth define STS Token Response
@@ -726,71 +723,6 @@ func importMutableParameters(params *diskVolumeArgs, mutable *ModifyParameters) 
 	}
 }
 
-func getMountedVolumeDevice(mnts []k8smount.MountInfo, targetPath string) string {
-	for _, mnt := range mnts {
-		if mnt.MountPoint == targetPath {
-			return mnt.Root
-		}
-	}
-	return ""
-}
-
-func isDeviceMountedAt(mnts []k8smount.MountInfo, device, targetPath string) bool {
-	for _, mnt := range mnts {
-		if mnt.MountPoint == targetPath && mnt.Source == device {
-			return true
-		}
-	}
-	return false
-}
-
-const mountInfoPath = "/proc/self/mountinfo"
-
-func CheckDeviceAvailable(devicePath, volumeID, targetPath string) error {
-	return checkDeviceAvailable(mountInfoPath, devicePath, volumeID, targetPath)
-}
-
-func checkDeviceAvailable(mountinfoPath, devicePath, volumeID, targetPath string) error {
-	if devicePath == "" {
-		return fmt.Errorf("devicePath is empty, cannot used for Volume")
-	}
-
-	mnts, err := k8smount.ParseMountInfo(mountinfoPath)
-	if err != nil {
-		return err
-	}
-
-	// TODO: remove in next version
-	// Since devicePath has been change to symlink, we will never run into the following logics.
-	// block volume
-	if devicePath == "devtmpfs" {
-		device := getMountedVolumeDevice(mnts, targetPath)
-		newVolumeID, err := GetVolumeIDByDevice(device)
-		if err != nil {
-			return nil
-		}
-		if newVolumeID != volumeID {
-			return fmt.Errorf("device [%s] associate with volumeID: [%s] rather than volumeID: [%s]", device, newVolumeID, volumeID)
-		}
-
-		return nil
-	}
-
-	if !utils.IsFileExisting(devicePath) {
-		return fmt.Errorf("devicePath(%s) is empty, cannot used for Volume", devicePath)
-	}
-
-	// check the device is used for system
-	if devicePath == "/dev/vda" || devicePath == "/dev/vda1" {
-		klog.Warningf("checkDeviceAvailable: devicePath(%s) may be system device: %s", devicePath, volumeID)
-	}
-
-	if isDeviceMountedAt(mnts, devicePath, utils.KubeletRootDir) {
-		return fmt.Errorf("devicePath(%s) is used as DataDisk for kubelet, cannot used for Volume", devicePath)
-	}
-	return nil
-}
-
 func appendDiskTypes(resp *ecs20140526.DescribeAvailableResourceResponse, zoneID string, types []string) []string {
 	if resp == nil || resp.Body == nil || resp.Body.AvailableZones == nil {
 		return types
@@ -1125,98 +1057,6 @@ func getVolumeCountFromLabelerAnnotation(node *v1.Node) (int, error) {
 		return 0, fmt.Errorf("invalid max-disk annotation: %w", err)
 	}
 	return n, nil
-}
-
-func getUnmanagedDiskCount(getNode func() (*v1.Node, error), ecs cloud.ECSv2Interface, instanceID, regionID string, dev utilsio.DiskLister) (int, error) {
-	// An attached disk is not managed by us if:
-	// 1. it is not in node.Status.VolumesInUse or node.Status.VolumesAttached; and
-	// 2. it does not have the xattr set.
-	// 1 may fail because the info in node.Status is removed before ControllerUnpublishVolume.
-	// 2 may fail because the disk may be just attached and not have the xattr set yet.
-	// Combine 1 and 2 to get the accurate "not managed" disk list.
-	// We should exclude these disks from available count.
-	// e.g. static/dynamic PVs are managed, OS disk or manually attached disks are not managed.
-
-	disks, err := listDiskXattrs(dev)
-	if err != nil {
-		return 0, fmt.Errorf("failed to list devices: %w", err)
-	}
-	managedDisks := sets.KeySet(disks)
-
-	// To ensure all the managed attachedDisks also present in managedDisks,
-	// ECS OpenAPI should goes after ListDisks because the just detached disk should
-	// disappear from ListDisks after OpenAPI;
-	// ECS OpenAPI should goes before getNode because the just attached disk should
-	// appear in node before OpenAPI;
-	attachedDisks, err := GetAttachedCloudDisks(ecs, instanceID, regionID)
-	if err != nil {
-		return 0, err
-	}
-	klog.Infof("getVolumeCount: found %d attached disks", len(attachedDisks))
-
-	node, err := getNode()
-	if err != nil {
-		return 0, err
-	}
-
-	for id := range CollectManagedVolumes(node) {
-		managedDisks.Insert(id)
-	}
-
-	unmanaged := 0
-	for _, disk := range attachedDisks {
-		if !managedDisks.Has(disk) {
-			klog.Infof("getVolumeCount: disk %s is not managed by us", disk)
-			unmanaged++
-		}
-	}
-
-	return unmanaged, nil
-}
-
-// checkRundVolumeExpand
-func checkRundVolumeExpand(req *csi.NodeExpandVolumeRequest) (bool, error) {
-	klog.Infof("checkRundVolumeExpand: volumePath: %s", req.VolumePath)
-	pvName := utils.GetPvNameFormPodMnt(req.VolumePath)
-	if pvName == "" {
-		klog.Errorf("checkRundVolumeExpand: cannot get pvname from volumePath %s", req.VolumePath)
-		return false, perrors.Errorf("cannot get pvname from volumePath %s for volume %s", req.VolumePath, req.VolumeId)
-	}
-	var grpcVolume string
-	socketFile := filepath.Join(RundSocketDir, pvName)
-	if utils.IsFileExisting(socketFile) {
-		grpcVolume = pvName
-	} else {
-		socketFile = filepath.Join(RundSocketDir, req.VolumeId)
-		if !utils.IsFileExisting(socketFile) {
-			klog.Infof("checkRundVolumeExpand: socketfile: %s not exists, trying runc expanding", socketFile)
-			return false, nil
-		}
-		grpcVolume = req.VolumeId
-	}
-	klog.Infof("checkRundVolumeExpand: rund socket dir: %s exists", socketFile)
-
-	// connect to rund server with timeout
-	clientConn, err := net.DialTimeout("unix", socketFile, 1*time.Second)
-	if err != nil {
-		klog.Errorf("checkRundExpand: volume %s, volumepath %s, connect to rund server with error: %s", req.VolumeId, req.VolumePath, err.Error())
-		return true, perrors.Errorf("checkRundExpand: volume %s, volumepath %s, connect to rund server with error: %s", req.VolumeId, req.VolumePath, err.Error())
-	}
-	defer clientConn.Close()
-
-	// send volume spec to rund to expand volume fs
-	volumeSize := strconv.FormatInt(req.GetCapacityRange().GetRequiredBytes(), 10)
-	client := proto.NewExtendedStatusClient(ttrpc.NewClient(clientConn))
-	resp, err := client.ExpandVolume(context.Background(), &proto.ExpandVolumeRequest{
-		Volume: grpcVolume,
-	})
-	if err != nil {
-		klog.Errorf("checkRundExpand: volume %s, volumepath %s, connect to rund server with error response: %s", req.VolumeId, req.VolumePath, err.Error())
-		return true, perrors.Errorf("checkRundExpand: volume %s, volumepath %s, connect to rund server with error response: %s", req.VolumeId, req.VolumePath, err.Error())
-	}
-
-	klog.Infof("RundVolumeExpand: Expand VolumeFS(%s) to(%s) successful with response: %s", pvName, volumeSize, resp)
-	return true, nil
 }
 
 func makeVolumeSnapshot(snapName string, snapContentName string) *volumeSnapshotV1.VolumeSnapshot {
