+++ pkg/disk/utils_test.go
@@ -1,5 +1,3 @@
-//go:build !windows
-
 /*
 Copyright 2019 The Kubernetes Authors.
 
@@ -20,11 +18,8 @@ package disk
 import (
 	"cmp"
 	"context"
-	"errors"
 	"fmt"
 	"os"
-	"path"
-	"path/filepath"
 	"testing"
 
 	ecs20140526 "github.com/alibabacloud-go/ecs-20140526/v7/client"
@@ -34,14 +29,12 @@ import (
 	"github.com/kubernetes-sigs/alibaba-cloud-csi-driver/pkg/cloud"
 	"github.com/kubernetes-sigs/alibaba-cloud-csi-driver/pkg/cloud/metadata"
 	"github.com/stretchr/testify/assert"
-	"golang.org/x/sys/unix"
 	corev1 "k8s.io/api/core/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/kubernetes/fake"
 	"k8s.io/client-go/tools/record"
 	"k8s.io/klog/v2/ktesting"
-	"k8s.io/mount-utils"
 	"k8s.io/utils/ptr"
 )
 
@@ -221,52 +214,6 @@ const DescribeDisksResponse = `{
 	}
 }`
 
-type MockDisks struct {
-	disks []string
-	base  string
-}
-
-func (m *MockDisks) ListDisks() ([]string, error) {
-	return m.disks, nil
-}
-
-func (m *MockDisks) AddDisk(t testing.TB, path string, diskID []byte) {
-	p := filepath.Join(m.base, path)
-	m.disks = append(m.disks, p)
-	assert.NoError(t, os.WriteFile(p, []byte{}, 0644))
-	if diskID != nil {
-		assert.NoError(t, unix.Setxattr(p, DiskXattrName, diskID, 0))
-	}
-}
-
-const longDiskID = "d-some-very-looooooooooooooooog-value-that-cause-getxattr-to-fail"
-
-func TestGetUnmanagedDiskCount(t *testing.T) {
-	testDiskXattr(t)
-
-	ctrl := gomock.NewController(t)
-	c := cloud.NewMockECSv2Interface(ctrl)
-
-	var describeDisksResp ecs20140526.DescribeDisksResponse
-	cloud.UnmarshalV2Response([]byte(DescribeDisksResponse), &describeDisksResp)
-	c.EXPECT().DescribeDisks(gomock.Any()).Return(&describeDisksResp, nil)
-
-	dev := MockDisks{base: t.TempDir() + "/dev"}
-	assert.NoError(t, os.MkdirAll(dev.base, 0755))
-
-	// add xattr to CSI attached disk
-	dev.AddDisk(t, "node-for-testingdetachingdisk", []byte("d-testingdetachingdisk"))
-	// manually attached disk has no xattr
-	dev.AddDisk(t, "node-for-2zeh74nnxxrobxz49eug", nil)
-	// an arbitrary error for getxattr, we should ignore it
-	dev.AddDisk(t, "node-for-testinglocaldisk", []byte(longDiskID))
-
-	getNode := func() (*corev1.Node, error) { return testNode(), nil }
-	count, err := getUnmanagedDiskCount(getNode, c, "i-2ze0yyw7rf00yz9fttpg", "cn-beijing", &dev)
-	assert.NoError(t, err)
-	assert.Equal(t, 2, count) // 1 system disk (d-2ze49fivxwkwxl36o1d3) + 1 manually attached (d-2zeh74nnxxrobxz49eug)
-}
-
 func TestGetAvailableDiskTypes(t *testing.T) {
 	descJson := `{
 	"RequestId": "6ECCECF5-945D-58FB-9BA9-312DBEE3F611",
@@ -442,86 +389,6 @@ func TestPatchForNodeExisting(t *testing.T) {
 	// When we cannot get diskTypes from OpenAPI, we should not remove existing labels
 	patch = patchForNode(node, 16, nil)
 	assert.Nil(t, patch)
-}
-
-func writeMountinfo(t *testing.T, mountInfo string) string {
-	mountInfoPath := path.Join(os.TempDir(), "mountinfo")
-	err := os.WriteFile(mountInfoPath, []byte(mountInfo), 0o644)
-	assert.NoError(t, err)
-	return mountInfoPath
-}
-
-func parseMountinfo(t *testing.T, mountInfo string) []mount.MountInfo {
-	mountInfoPath := writeMountinfo(t, mountInfo)
-	mnts, err := mount.ParseMountInfo(mountInfoPath)
-	assert.NoError(t, err)
-	return mnts
-}
-
-func TestGetMountedVolumeDevice(t *testing.T) {
-	cases := []struct {
-		name      string
-		mountInfo []mount.MountInfo
-		device    string
-	}{
-		{
-			name:      "mounted",
-			mountInfo: parseMountinfo(t, "707 97 0:5 /vdc /path/to/volumeDevice rw,nosuid shared:21 - devtmpfs devtmpfs rw,size=7901960k,nr_inodes=1975490,mode=755"),
-			device:    "/vdc",
-		},
-		{
-			name:      "not mounted",
-			mountInfo: parseMountinfo(t, "707 97 0:5 /vdc /path/to/another_volumeDevice rw,nosuid shared:21 - devtmpfs devtmpfs rw,size=7901960k,nr_inodes=1975490,mode=755"),
-			device:    "",
-		},
-	}
-
-	for _, test := range cases {
-		t.Run(test.name, func(t *testing.T) {
-			t.Parallel()
-			device := getMountedVolumeDevice(test.mountInfo, "/path/to/volumeDevice")
-			assert.Equal(t, test.device, device)
-		})
-	}
-}
-
-func TestIsDeviceMountedAt(t *testing.T) {
-	cases := []struct {
-		name      string
-		mountInfo []mount.MountInfo
-		mounted   bool
-	}{
-		{
-			name:      "mounted",
-			mountInfo: parseMountinfo(t, "291 97 253:16 / /path/to/mountpoint rw,relatime shared:160 - ext4 /dev/vdb rw"),
-			mounted:   true,
-		},
-		{
-			name:      "wrong device",
-			mountInfo: parseMountinfo(t, "291 97 253:16 / /path/to/mountpoint rw,relatime shared:160 - ext4 /dev/vdc rw"),
-			mounted:   false,
-		},
-		{
-			name:      "wrong path",
-			mountInfo: parseMountinfo(t, "291 97 253:16 / /path/to/another/mountpoint rw,relatime shared:160 - ext4 /dev/vdb rw"),
-			mounted:   false,
-		},
-	}
-
-	for _, test := range cases {
-		t.Run(test.name, func(t *testing.T) {
-			t.Parallel()
-			mounted := isDeviceMountedAt(test.mountInfo, "/dev/vdb", "/path/to/mountpoint")
-			assert.Equal(t, test.mounted, mounted)
-		})
-	}
-}
-
-func TestCheckDeviceAvailableError(t *testing.T) {
-	err := checkDeviceAvailable("/not/exist", "/dev/vdc", "d-2zedmdfyiz2num45yx60", "/path/to/mountpoint")
-	if !errors.Is(err, os.ErrNotExist) {
-		t.Errorf("expected os.ErrNotExist, got %v", err)
-	}
 }
 
 func TestDiskSize(t *testing.T) {
