+++ pkg/utils/mounter_unix.go
@@ -1,3 +1,5 @@
+//go:build unix
+
 package utils
 
 import (
